♻️ refactor: 공통 유틸 이동 - application/util -> common/utils - 경로 수정으로 추가적인 변경 반영
