+++ src/application/hooks/api/tags/index.ts
@@ -3,7 +3,7 @@ import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 
 import { useSuspendedQuery } from "@/application/hooks/api/core";
 import type { QuerySelectOption } from "@/application/hooks/api/core/types";
-import { delay } from "@/application/util";
+import { delay } from "@/common/utils";
 import { api } from "@/infra/api";
 import type {
   GetFavoriteTagsResponse,
