+++ src/common/components/DropDown/DropDown.tsx
@@ -9,8 +9,8 @@ import type {
 import { createContext, useContext, useState } from "react";
 import { css } from "twin.macro";
 
-import { withDelay } from "@/application/util/delay";
 import { useClickOutside } from "@/common/hooks";
+import { withDelay } from "@/common/utils";
 
 const DropDownContext = createContext(false);
 const DropDownSetContext = createContext<Dispatch<SetStateAction<boolean>>>(() => null);
