+++ src/common/utils/animation.ts
