+++ src/common/utils/device.ts
