+++ src/common/utils/image-util.ts
