+++ src/components/meme/MemeInfo/DropDown/MemeExport.tsx
@@ -6,10 +6,10 @@ import {
   useMemeDetailById,
   usePostMemeToSharedCollection,
 } from "@/application/hooks";
-import { DOMAIN, PATH } from "@/application/util";
 import { DropDown } from "@/common/components/DropDown";
 import { Icon } from "@/common/components/Icon";
 import { useToast } from "@/common/hooks";
+import { DOMAIN, PATH } from "@/common/utils";
 
 interface Props {
   id: string;
