+++ src/components/meme/MemeInfo/Modal/MemeShareModal.tsx
@@ -1,9 +1,9 @@
 import { useAuth, useMemeDetailById } from "@/application/hooks";
 import { usePostMemeToSharedCollection } from "@/application/hooks/api/collection";
-import { DOMAIN, PATH } from "@/application/util";
 import { Modal } from "@/common/components/Modal";
 import { Photo } from "@/common/components/Photo";
 import { useToast } from "@/common/hooks";
+import { DOMAIN, PATH } from "@/common/utils";
 import {
   ClipboardCopyButton,
   KakaoShareButton,
