+++ src/components/meme/MemeItem/MemeItem.tsx
@@ -1,10 +1,10 @@
 import { memo } from "react";
 
 import { useMoveMemeDetail } from "@/application/hooks";
-import { isEncodingError } from "@/application/util";
 import { Icon } from "@/common/components/Icon";
 import { Photo } from "@/common/components/Photo";
 import { useOverlay } from "@/common/hooks";
+import { isEncodingError } from "@/common/utils";
 import type { Meme } from "@/types";
 
 import { MemeActionSheet } from "../ActionSheet";
