+++ src/components/search/SearchResult/SearchResultList.tsx
@@ -2,8 +2,8 @@ import Link from "next/link";
 
 import type { RecentSearch } from "@/application/hooks";
 import { useGetTagSearch } from "@/application/hooks";
-import { PATH } from "@/application/util";
 import { Icon } from "@/common/components/Icon";
+import { PATH } from "@/common/utils";
 import { SearchItem } from "@/components/search";
 
 interface Prop {
