+++ src/components/tags/TagCategory/FavoriteCategory.tsx
@@ -2,10 +2,10 @@ import { Content, Header, Item, Trigger } from "@radix-ui/react-accordion";
 import { useRouter } from "next/router";
 
 import { useAuth, useDeleteFavoriteTag, useGetFavoriteTags } from "@/application/hooks";
-import { PATH } from "@/application/util";
 import { Icon } from "@/common/components/Icon";
 import { Photo } from "@/common/components/Photo";
 import { useToast } from "@/common/hooks";
+import { PATH } from "@/common/utils";
 
 import { CategoryTitle } from "./CategoryTitle";
 import { useTagCategoryContext } from "./context";
