+++ src/pages/index.tsx
@@ -1,10 +1,10 @@
 import type { NextPage } from "next";
 
-import { DEFAULT_DESCRIPTION, SITE_NAME } from "@/application/util";
 import { IntroPageNavigation } from "@/common/components/Navigation";
 import type { NextSeoProps } from "@/common/components/NextSeo";
 import { NextSeo } from "@/common/components/NextSeo";
 import { PullToRefresh } from "@/common/components/PullToRefresh";
+import { DEFAULT_DESCRIPTION, SITE_NAME } from "@/common/utils";
 import { MemeListContainer } from "@/components/home";
 
 const HomePage: NextPage = () => {
