+++ src/pages/search/index.tsx
@@ -3,12 +3,12 @@ import { useRouter } from "next/router";
 import { Suspense } from "react";
 
 import { useRecentSearch } from "@/application/hooks";
-import { DEFAULT_DESCRIPTION, PATH, SITE_NAME } from "@/application/util";
 import { SearchPageNavigation } from "@/common/components/Navigation";
 import type { NextSeoProps } from "@/common/components/NextSeo";
 import { NextSeo } from "@/common/components/NextSeo";
 import { SSRSuspense } from "@/common/components/Suspense";
 import { useDebounce, useInput } from "@/common/hooks";
+import { DEFAULT_DESCRIPTION, PATH, SITE_NAME } from "@/common/utils";
 import {
   SearchInput,
   SearchPopularList,
